+++ src/pulse/TxDriver.h
@@ -21,7 +21,7 @@ class TxDriver {
   virtual void setFramingMode(FramingMode mode) = 0;
   virtual int write(uint8_t byte) = 0;
   virtual size_t write(const uint8_t* data, size_t len) = 0;
-  virtual bool begin(uint16_t bitFrequencyHz) = 0;
+  virtual bool begin(uint32_t bitFrequencyHz) = 0;
   virtual void flush() {};
   virtual void end() {};
 };
